backend/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
imbalanced-learn==0.8.0
imblearn==0.0
joblib==1.0.1
numpy==1.17.4
pandas==0.25.3
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
threadpoolctl==2.1.0
